refactor(mm-next): move getPostsAndPostscountFromGqlData
