+++ packages/mirror-media-next/pages/section/[slug].js
@@ -3,11 +3,11 @@ import dynamic from 'next/dynamic'
 
 import SectionArticles from '../../components/shared/section-articles'
 import { ENV } from '../../config/index.mjs'
+import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api'
 import {
-  fetchHeaderDataInDefaultPageLayout,
+  getSectionAndTopicFromDefaultHeaderData,
   getPostsAndPostscountFromGqlData,
-} from '../../utils/api'
-import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/data-process'
+} from '../../utils/data-process'
 import { getLogTraceObject } from '../../utils'
 import {
   handleAxiosResponse,
