+++ packages/mirror-media-next/utils/data-process.js
@@ -41,7 +41,24 @@ const getSectionFromPremiumHeaderData = (headerData) => {
   return sectionData
 }
 
+/**
+ * @template T
+ * @param {import('@apollo/client').ApolloQueryResult<any> | undefined} gqlData
+ * @returns {[number, T[]]}
+ */
+const getPostsAndPostscountFromGqlData = (gqlData) => {
+  if (!gqlData) {
+    return [0, []]
+  }
+
+  const data = gqlData.data
+  const postsCount = data?.postsCount || 0
+  const posts = data?.posts || []
+  return [postsCount, posts]
+}
+
 export {
   getSectionAndTopicFromDefaultHeaderData,
   getSectionFromPremiumHeaderData,
+  getPostsAndPostscountFromGqlData,
 }
